requirements.txt
Flask
Flask-SQLAlchemy
Flasgger
python-dotenv
